Fix class name for update body and title
