+++ src/handlers/me/articles/drafts/body/update/me_articles_drafts_body_update.py
@@ -8,7 +8,7 @@ from user_util import UserUtil
 from text_sanitizer import TextSanitizer
 
 
-class MeArticlesPublicUpdateBody(LambdaBase):
+class MeArticlesDraftsBodyUpdate(LambdaBase):
     def get_schema(self):
         return {
             'type': 'object',
@@ -22,32 +22,34 @@ class MeArticlesPublicUpdateBody(LambdaBase):
     def validate_params(self):
         UserUtil.verified_phone_and_email(self.event)
         validate(self.params, self.get_schema(), format_checker=FormatChecker())
-
         DBUtil.validate_article_existence(
             self.dynamodb,
             self.params['article_id'],
             user_id=self.event['requestContext']['authorizer']['claims']['cognito:username'],
-            status='public',
+            status='draft',
             version=2
         )
 
     def exec_main_proc(self):
-        article_content_edit_table = self.dynamodb.Table(os.environ['ARTICLE_CONTENT_EDIT_TABLE_NAME'])
-
-        expression_attribute_values = {
-            ':user_id': self.event['requestContext']['authorizer']['claims']['cognito:username'],
-            ':body': TextSanitizer.sanitize_article_body_v2(self.params.get('body'))
-        }
-        DBUtil.items_values_empty_to_none(expression_attribute_values)
-
-        article_content_edit_table.update_item(
-            Key={
-                'article_id': self.params['article_id'],
-            },
-            UpdateExpression="set user_id=:user_id, body=:body",
-            ExpressionAttributeValues=expression_attribute_values
-        )
+        self.__update_article_content()
 
         return {
             'statusCode': 200
         }
+
+    def __update_article_content(self):
+        article_content_table = self.dynamodb.Table(os.environ['ARTICLE_CONTENT_TABLE_NAME'])
+
+        expression_attribute_values = {
+            ':body': TextSanitizer.sanitize_article_body_v2(self.params.get('body'))
+        }
+
+        DBUtil.items_values_empty_to_none(expression_attribute_values)
+
+        article_content_table.update_item(
+            Key={
+                'article_id': self.params['article_id'],
+            },
+            UpdateExpression="set body=:body",
+            ExpressionAttributeValues=expression_attribute_values
+        )
