+++ src/handlers/me/articles/drafts/title/update/me_articles_drafts_title_update.py
@@ -3,20 +3,20 @@ import os
 import settings
 from lambda_base import LambdaBase
 from jsonschema import validate, FormatChecker
+from text_sanitizer import TextSanitizer
 from db_util import DBUtil
 from user_util import UserUtil
-from text_sanitizer import TextSanitizer
 
 
-class MeArticlesDraftsUpdateBody(LambdaBase):
+class MeArticlesDraftsTitleUpdate(LambdaBase):
     def get_schema(self):
         return {
             'type': 'object',
             'properties': {
                 'article_id': settings.parameters['article_id'],
-                'body': settings.parameters['body']
+                'title': settings.parameters['title']
             },
-            'required': ['article_id', 'body']
+            'required': ['article_id', 'title']
         }
 
     def validate_params(self):
@@ -31,6 +31,7 @@ class MeArticlesDraftsUpdateBody(LambdaBase):
         )
 
     def exec_main_proc(self):
+        self.__update_article_info()
         self.__update_article_content()
 
         return {
@@ -39,9 +40,8 @@ class MeArticlesDraftsUpdateBody(LambdaBase):
 
     def __update_article_content(self):
         article_content_table = self.dynamodb.Table(os.environ['ARTICLE_CONTENT_TABLE_NAME'])
-
         expression_attribute_values = {
-            ':body': TextSanitizer.sanitize_article_body_v2(self.params.get('body'))
+            ':title': TextSanitizer.sanitize_text(self.params.get('title'))
         }
 
         DBUtil.items_values_empty_to_none(expression_attribute_values)
@@ -50,6 +50,21 @@ class MeArticlesDraftsUpdateBody(LambdaBase):
             Key={
                 'article_id': self.params['article_id'],
             },
-            UpdateExpression="set body=:body",
+            UpdateExpression="set title=:title",
+            ExpressionAttributeValues=expression_attribute_values
+        )
+
+    def __update_article_info(self):
+        article_info_table = self.dynamodb.Table(os.environ['ARTICLE_INFO_TABLE_NAME'])
+        expression_attribute_values = {
+            ':title': TextSanitizer.sanitize_text(self.params.get('title')),
+        }
+        DBUtil.items_values_empty_to_none(expression_attribute_values)
+
+        article_info_table.update_item(
+            Key={
+                'article_id': self.params['article_id'],
+            },
+            UpdateExpression="set title = :title",
             ExpressionAttributeValues=expression_attribute_values
         )
