+++ src/handlers/me/articles/public/body/update/me_articles_public_body_update.py
@@ -3,20 +3,20 @@ import os
 import settings
 from lambda_base import LambdaBase
 from jsonschema import validate, FormatChecker
-from text_sanitizer import TextSanitizer
 from db_util import DBUtil
 from user_util import UserUtil
+from text_sanitizer import TextSanitizer
 
 
-class MeArticlesPublicUpdateTitle(LambdaBase):
+class MeArticlesPublicBodyUpdate(LambdaBase):
     def get_schema(self):
         return {
             'type': 'object',
             'properties': {
                 'article_id': settings.parameters['article_id'],
-                'title': settings.parameters['title']
+                'body': settings.parameters['body']
             },
-            'required': ['article_id', 'title']
+            'required': ['article_id', 'body']
         }
 
     def validate_params(self):
@@ -36,7 +36,7 @@ class MeArticlesPublicUpdateTitle(LambdaBase):
 
         expression_attribute_values = {
             ':user_id': self.event['requestContext']['authorizer']['claims']['cognito:username'],
-            ':title': TextSanitizer.sanitize_text(self.params.get('title'))
+            ':body': TextSanitizer.sanitize_article_body_v2(self.params.get('body'))
         }
         DBUtil.items_values_empty_to_none(expression_attribute_values)
 
@@ -44,7 +44,7 @@ class MeArticlesPublicUpdateTitle(LambdaBase):
             Key={
                 'article_id': self.params['article_id'],
             },
-            UpdateExpression="set user_id=:user_id, title=:title",
+            UpdateExpression="set user_id=:user_id, body=:body",
             ExpressionAttributeValues=expression_attribute_values
         )
