+++ src/handlers/me/articles/public/title/update/me_articles_public_title_update.py
@@ -8,7 +8,7 @@ from db_util import DBUtil
 from user_util import UserUtil
 
 
-class MeArticlesDraftsUpdateTitle(LambdaBase):
+class MeArticlesPublicTitleUpdate(LambdaBase):
     def get_schema(self):
         return {
             'type': 'object',
@@ -22,49 +22,32 @@ class MeArticlesDraftsUpdateTitle(LambdaBase):
     def validate_params(self):
         UserUtil.verified_phone_and_email(self.event)
         validate(self.params, self.get_schema(), format_checker=FormatChecker())
+
         DBUtil.validate_article_existence(
             self.dynamodb,
             self.params['article_id'],
             user_id=self.event['requestContext']['authorizer']['claims']['cognito:username'],
-            status='draft',
+            status='public',
             version=2
         )
 
     def exec_main_proc(self):
-        self.__update_article_info()
-        self.__update_article_content()
+        article_content_edit_table = self.dynamodb.Table(os.environ['ARTICLE_CONTENT_EDIT_TABLE_NAME'])
+
+        expression_attribute_values = {
+            ':user_id': self.event['requestContext']['authorizer']['claims']['cognito:username'],
+            ':title': TextSanitizer.sanitize_text(self.params.get('title'))
+        }
+        DBUtil.items_values_empty_to_none(expression_attribute_values)
+
+        article_content_edit_table.update_item(
+            Key={
+                'article_id': self.params['article_id'],
+            },
+            UpdateExpression="set user_id=:user_id, title=:title",
+            ExpressionAttributeValues=expression_attribute_values
+        )
 
         return {
             'statusCode': 200
         }
-
-    def __update_article_content(self):
-        article_content_table = self.dynamodb.Table(os.environ['ARTICLE_CONTENT_TABLE_NAME'])
-        expression_attribute_values = {
-            ':title': TextSanitizer.sanitize_text(self.params.get('title'))
-        }
-
-        DBUtil.items_values_empty_to_none(expression_attribute_values)
-
-        article_content_table.update_item(
-            Key={
-                'article_id': self.params['article_id'],
-            },
-            UpdateExpression="set title=:title",
-            ExpressionAttributeValues=expression_attribute_values
-        )
-
-    def __update_article_info(self):
-        article_info_table = self.dynamodb.Table(os.environ['ARTICLE_INFO_TABLE_NAME'])
-        expression_attribute_values = {
-            ':title': TextSanitizer.sanitize_text(self.params.get('title')),
-        }
-        DBUtil.items_values_empty_to_none(expression_attribute_values)
-
-        article_info_table.update_item(
-            Key={
-                'article_id': self.params['article_id'],
-            },
-            UpdateExpression="set title = :title",
-            ExpressionAttributeValues=expression_attribute_values
-        )
