+++ tests/handlers/me/articles/drafts/title/update/test_me_articles_drafts_title_update.py
@@ -1,12 +1,12 @@
 import json
 import os
 from unittest import TestCase
-from me_articles_drafts_update_body import MeArticlesDraftsUpdateBody
+from me_articles_drafts_title_update import MeArticlesDraftsTitleUpdate
 from unittest.mock import patch, MagicMock
 from tests_util import TestsUtil
 
 
-class TestMeArticlesDraftsUpdateBody(TestCase):
+class TestMeArticlesDraftsTitleUpdate(TestCase):
     dynamodb = TestsUtil.get_dynamodb_client()
 
     @classmethod
@@ -72,8 +72,8 @@ class TestMeArticlesDraftsUpdateBody(TestCase):
         TestsUtil.delete_all_tables(self.dynamodb)
 
     def assert_bad_request(self, params):
-        me_articles_drafts_update_body = MeArticlesDraftsUpdateBody(params, {}, self.dynamodb)
-        response = me_articles_drafts_update_body.main()
+        me_articles_drafts_title_update = MeArticlesDraftsTitleUpdate(params, {}, self.dynamodb)
+        response = me_articles_drafts_title_update.main()
 
         self.assertEqual(response['statusCode'], 400)
 
@@ -83,7 +83,7 @@ class TestMeArticlesDraftsUpdateBody(TestCase):
                 'article_id': 'draftId00001'
             },
             'body': {
-                'body': '<p>update body</p>',
+                'title': 'update title'
             },
             'requestContext': {
                 'authorizer': {
@@ -98,17 +98,25 @@ class TestMeArticlesDraftsUpdateBody(TestCase):
 
         params['body'] = json.dumps(params['body'])
 
+        article_info_before = self.article_info_table.scan()['Items']
         article_content_before = self.article_content_table.scan()['Items']
 
-        me_articles_drafts_update_body = MeArticlesDraftsUpdateBody(params, {}, self.dynamodb)
-        response = me_articles_drafts_update_body.main()
+        me_articles_drafts_update = MeArticlesDraftsTitleUpdate(params, {}, self.dynamodb)
 
+        response = me_articles_drafts_update.main()
+
+        article_info_after = self.article_info_table.scan()['Items']
         article_content_after = self.article_content_table.scan()['Items']
 
         self.assertEqual(response['statusCode'], 200)
+        self.assertEqual(len(article_info_after) - len(article_info_before), 0)
         self.assertEqual(len(article_content_after) - len(article_content_before), 0)
 
+        self.assertEqual(params['requestContext']['authorizer']['claims']['cognito:username'],
+                         article_info_after[0]['user_id'])
+
         for key, value in json.loads(params['body']).items():
+            self.assertEqual(value, article_info_after[0][key])
             self.assertEqual(value, article_content_after[0][key])
 
     def test_main_ok_with_empty_string(self):
@@ -117,7 +125,7 @@ class TestMeArticlesDraftsUpdateBody(TestCase):
                 'article_id': 'draftId00001'
             },
             'body': {
-                'body': ''
+                'title': ''
             },
             'requestContext': {
                 'authorizer': {
@@ -132,13 +140,18 @@ class TestMeArticlesDraftsUpdateBody(TestCase):
 
         params['body'] = json.dumps(params['body'])
 
-        response = MeArticlesDraftsUpdateBody(params, {}, self.dynamodb).main()
+        response = MeArticlesDraftsTitleUpdate(params, {}, self.dynamodb).main()
+        article_info = self.article_info_table.get_item(
+            Key={'article_id': params['pathParameters']['article_id']}
+        )['Item']
+
         article_content = self.article_content_table.get_item(
             Key={'article_id': params['pathParameters']['article_id']}
         )['Item']
 
         self.assertEqual(response['statusCode'], 200)
-        self.assertEqual(article_content['body'], None)
+        self.assertEqual(article_info['title'], None)
+        self.assertEqual(article_content['title'], None)
 
     def test_main_ok_article_content_not_exists(self):
         params = {
@@ -146,7 +159,7 @@ class TestMeArticlesDraftsUpdateBody(TestCase):
                 'article_id': 'draftId00003'
             },
             'body': {
-                'body': 'A' * 65535,
+                'title': 'A' * 255
             },
             'requestContext': {
                 'authorizer': {
@@ -161,19 +174,28 @@ class TestMeArticlesDraftsUpdateBody(TestCase):
 
         params['body'] = json.dumps(params['body'])
 
+        article_info_before = self.article_info_table.scan()['Items']
         article_content_before = self.article_content_table.scan()['Items']
 
-        me_articles_drafts_update_body = MeArticlesDraftsUpdateBody(params, {}, self.dynamodb)
-        response = me_articles_drafts_update_body.main()
+        me_articles_drafts_update = MeArticlesDraftsTitleUpdate(params, {}, self.dynamodb)
 
+        response = me_articles_drafts_update.main()
+
+        article_info_after = self.article_info_table.scan()['Items']
         article_content_after = self.article_content_table.scan()['Items']
 
         self.assertEqual(response['statusCode'], 200)
+        self.assertEqual(len(article_info_after) - len(article_info_before), 0)
         self.assertEqual(len(article_content_after) - len(article_content_before), 1)
 
+        self.assertEqual(params['requestContext']['authorizer']['claims']['cognito:username'],
+                         article_info_after[0]['user_id'])
+
         for key, value in json.loads(params['body']).items():
-            target = [content for content in article_content_after if content['article_id'] == 'draftId00003'][0]
-            self.assertEqual(value, target[key])
+            target_info = [info for info in article_info_after if info['article_id'] == 'draftId00003'][0]
+            self.assertEqual(value, target_info[key])
+            target_content = [content for content in article_content_after if content['article_id'] == 'draftId00003'][0]
+            self.assertEqual(value, target_content[key])
 
     def test_call_validate_article_existence(self):
         params = {
@@ -181,7 +203,8 @@ class TestMeArticlesDraftsUpdateBody(TestCase):
                 'article_id': 'draftId00001'
             },
             'body': {
-                'body': '<p>update body</p>'
+                'eye_catch_url': 'http://example.com/update',
+                'title': 'update title'
             },
             'requestContext': {
                 'authorizer': {
@@ -197,8 +220,8 @@ class TestMeArticlesDraftsUpdateBody(TestCase):
         params['body'] = json.dumps(params['body'])
 
         mock_lib = MagicMock()
-        with patch('me_articles_drafts_update_body.DBUtil', mock_lib):
-            MeArticlesDraftsUpdateBody(params, {}, self.dynamodb).main()
+        with patch('me_articles_drafts_title_update.DBUtil', mock_lib):
+            MeArticlesDraftsTitleUpdate(params, {}, self.dynamodb).main()
             args, kwargs = mock_lib.validate_article_existence.call_args
 
             self.assertTrue(mock_lib.validate_article_existence.called)
@@ -208,14 +231,14 @@ class TestMeArticlesDraftsUpdateBody(TestCase):
             self.assertEqual(kwargs['status'], 'draft')
             self.assertEqual(kwargs['version'], 2)
 
-    def test_call_sanitize_article_body_v2(self):
-        body_str = '<p>update body</p>'
+    def test_call_sanitize_text(self):
+        title_str = 'test_title'
         params = {
             'pathParameters': {
                 'article_id': 'draftId00001'
             },
             'body': {
-                'body': body_str
+                'title': title_str
             },
             'requestContext': {
                 'authorizer': {
@@ -231,11 +254,11 @@ class TestMeArticlesDraftsUpdateBody(TestCase):
         params['body'] = json.dumps(params['body'])
 
         mock_lib = MagicMock()
-        with patch('me_articles_drafts_update_body.TextSanitizer', mock_lib):
-            MeArticlesDraftsUpdateBody(params, {}, self.dynamodb).main()
-            args, kwargs = mock_lib.sanitize_article_body_v2.call_args
-            self.assertTrue(mock_lib.sanitize_article_body_v2.called)
-            self.assertEqual(args[0], body_str)
+        with patch('me_articles_drafts_title_update.TextSanitizer', mock_lib):
+            MeArticlesDraftsTitleUpdate(params, {}, self.dynamodb).main()
+            args, kwargs = mock_lib.sanitize_text.call_args
+            self.assertTrue(mock_lib.sanitize_text.called)
+            self.assertEqual(args[0], title_str)
 
     def test_validation_with_no_params(self):
         params = {}
@@ -257,7 +280,7 @@ class TestMeArticlesDraftsUpdateBody(TestCase):
     def test_validation_with_no_path_params(self):
         params = {
             'body': {
-                'body': 'A' * 200
+                'title': 'A' * 200
             },
             'pathParameters': {}
         }
@@ -266,10 +289,10 @@ class TestMeArticlesDraftsUpdateBody(TestCase):
 
         self.assert_bad_request(params)
 
-    def test_validation_body_max(self):
+    def test_validation_title_max(self):
         params = {
             'body': {
-                'body': 'A' * 65536
+                'title': 'A' * 256
             },
             'pathParameters': {
                 'article_id': 'A' * 12
@@ -283,7 +306,7 @@ class TestMeArticlesDraftsUpdateBody(TestCase):
     def test_validation_article_id_max(self):
         params = {
             'body': {
-                'body': 'A' * 50
+                'title': 'A' * 50
             },
             'pathParameters': {
                 'article_id': 'A' * 13
@@ -297,7 +320,7 @@ class TestMeArticlesDraftsUpdateBody(TestCase):
     def test_validation_article_id_min(self):
         params = {
             'body': {
-                'body': 'A' * 50
+                'title': 'A' * 50
             },
             'pathParameters': {
                 'article_id': 'A' * 11
