+++ tests/handlers/me/articles/public/body/update/test_me_articles_public_body_update.py
@@ -1,12 +1,12 @@
 import json
 import os
 from unittest import TestCase
-from me_articles_public_update_title import MeArticlesPublicUpdateTitle
+from me_articles_public_body_update import MeArticlesPublicBodyUpdate
 from unittest.mock import patch, MagicMock
 from tests_util import TestsUtil
 
 
-class TestMeArticlesPublicUpdate(TestCase):
+class TestMeArticlesPublicBodyUpdate(TestCase):
     dynamodb = TestsUtil.get_dynamodb_client()
 
     @classmethod
@@ -73,8 +73,8 @@ class TestMeArticlesPublicUpdate(TestCase):
         TestsUtil.delete_all_tables(self.dynamodb)
 
     def assert_bad_request(self, params):
-        me_articles_public_update_title = MeArticlesPublicUpdateTitle(params, {}, self.dynamodb)
-        response = me_articles_public_update_title.main()
+        me_articles_public_body_update = MeArticlesPublicBodyUpdate(params, {}, self.dynamodb)
+        response = me_articles_public_body_update.main()
 
         self.assertEqual(response['statusCode'], 400)
 
@@ -84,7 +84,7 @@ class TestMeArticlesPublicUpdate(TestCase):
                 'article_id': 'publicId0001'
             },
             'body': {
-                'title': 'title text'
+                'body': '<p>update body</p>'
             },
             'requestContext': {
                 'authorizer': {
@@ -100,7 +100,7 @@ class TestMeArticlesPublicUpdate(TestCase):
         article_content_edit_before = self.article_content_edit_table.scan()['Items']
 
         params['body'] = json.dumps(params['body'])
-        response = MeArticlesPublicUpdateTitle(params, {}, self.dynamodb).main()
+        response = MeArticlesPublicBodyUpdate(params, {}, self.dynamodb).main()
 
         article_content_edit_after = self.article_content_edit_table.scan()['Items']
 
@@ -123,7 +123,7 @@ class TestMeArticlesPublicUpdate(TestCase):
                 'article_id': 'publicId0001'
             },
             'body': {
-                'title': ''
+                'body': ''
             },
             'requestContext': {
                 'authorizer': {
@@ -137,7 +137,7 @@ class TestMeArticlesPublicUpdate(TestCase):
         }
 
         params['body'] = json.dumps(params['body'])
-        response = MeArticlesPublicUpdateTitle(params, {}, self.dynamodb).main()
+        response = MeArticlesPublicBodyUpdate(params, {}, self.dynamodb).main()
         article_content_edit = self.article_content_edit_table.get_item(
             Key={'article_id': params['pathParameters']['article_id']}
         )['Item']
@@ -147,11 +147,12 @@ class TestMeArticlesPublicUpdate(TestCase):
         expected_items = {
             'article_id': self.article_content_edit_items[0].get('article_id'),
             'eye_catch_url': self.article_content_edit_items[0].get('eye_catch_url'),
-            'title': None,
-            'body': self.article_content_edit_items[0].get('body'),
+            'title': self.article_content_edit_items[0].get('title'),
+            'body': None,
             'overview': self.article_content_edit_items[0].get('overview'),
             'user_id': self.article_content_edit_items[0].get('user_id'),
         }
+
         self.assertEqual(article_content_edit, expected_items)
 
     def test_main_ok_with_no_article_edit_content(self):
@@ -160,7 +161,7 @@ class TestMeArticlesPublicUpdate(TestCase):
                 'article_id': 'publicId0002'
             },
             'body': {
-                'title': 'A' * 255
+                'body': 'A' * 65535
             },
             'requestContext': {
                 'authorizer': {
@@ -176,7 +177,7 @@ class TestMeArticlesPublicUpdate(TestCase):
         article_content_edit_before = self.article_content_edit_table.scan()['Items']
 
         params['body'] = json.dumps(params['body'])
-        response = MeArticlesPublicUpdateTitle(params, {}, self.dynamodb).main()
+        response = MeArticlesPublicBodyUpdate(params, {}, self.dynamodb).main()
 
         article_content_edit_after = self.article_content_edit_table.scan()['Items']
 
@@ -199,7 +200,7 @@ class TestMeArticlesPublicUpdate(TestCase):
                 'article_id': 'publicId0001'
             },
             'body': {
-                'title': 'test text'
+                'body': '<p>update body</p>'
             },
             'requestContext': {
                 'authorizer': {
@@ -215,9 +216,10 @@ class TestMeArticlesPublicUpdate(TestCase):
         params['body'] = json.dumps(params['body'])
 
         mock_lib = MagicMock()
-        with patch('me_articles_public_update_title.DBUtil', mock_lib):
-            MeArticlesPublicUpdateTitle(params, {}, self.dynamodb).main()
+        with patch('me_articles_public_body_update.DBUtil', mock_lib):
+            MeArticlesPublicBodyUpdate(params, {}, self.dynamodb).main()
             args, kwargs = mock_lib.validate_article_existence.call_args
+
             self.assertTrue(mock_lib.validate_article_existence.called)
             self.assertTrue(args[0])
             self.assertTrue(args[1])
@@ -225,14 +227,14 @@ class TestMeArticlesPublicUpdate(TestCase):
             self.assertEqual(kwargs['status'], 'public')
             self.assertEqual(kwargs['version'], 2)
 
-    def test_call_sanitize_text(self):
-        title_str = 'test text'
+    def test_call_sanitize_article_body_v2(self):
+        body_str = '<p>update body</p>'
         params = {
             'pathParameters': {
                 'article_id': 'publicId0001'
             },
             'body': {
-                'title': title_str
+                'body': body_str
             },
             'requestContext': {
                 'authorizer': {
@@ -248,11 +250,11 @@ class TestMeArticlesPublicUpdate(TestCase):
         params['body'] = json.dumps(params['body'])
 
         mock_lib = MagicMock()
-        with patch('me_articles_public_update_title.TextSanitizer', mock_lib):
-            MeArticlesPublicUpdateTitle(params, {}, self.dynamodb).main()
-            args, kwargs = mock_lib.sanitize_text.call_args
-            self.assertTrue(mock_lib.sanitize_text.called)
-            self.assertEqual(args[0], title_str)
+        with patch('me_articles_public_body_update.TextSanitizer', mock_lib):
+            MeArticlesPublicBodyUpdate(params, {}, self.dynamodb).main()
+            args, kwargs = mock_lib.sanitize_article_body_v2.call_args
+            self.assertTrue(mock_lib.sanitize_article_body_v2.called)
+            self.assertEqual(args[0], body_str)
 
     def test_validation_with_no_params(self):
         params = {}
@@ -274,7 +276,7 @@ class TestMeArticlesPublicUpdate(TestCase):
     def test_validation_with_no_path_params(self):
         params = {
             'body': {
-                'title': 'A' * 200
+                'body': 'A' * 200
             },
             'pathParameters': {}
         }
@@ -283,10 +285,10 @@ class TestMeArticlesPublicUpdate(TestCase):
 
         self.assert_bad_request(params)
 
-    def test_validation_title_max(self):
+    def test_validation_body_max(self):
         params = {
             'body': {
-                'title': 'A' * 256
+                'body': 'A' * 65536
             },
             'pathParameters': {
                 'article_id': 'A' * 12
@@ -300,7 +302,7 @@ class TestMeArticlesPublicUpdate(TestCase):
     def test_validation_article_id_max(self):
         params = {
             'body': {
-                'title': 'A' * 50
+                'body': 'A' * 50
             },
             'pathParameters': {
                 'article_id': 'A' * 13
@@ -314,7 +316,7 @@ class TestMeArticlesPublicUpdate(TestCase):
     def test_validation_article_id_min(self):
         params = {
             'body': {
-                'title': 'A' * 50
+                'overview': 'A' * 50
             },
             'pathParameters': {
                 'article_id': 'A' * 11
